NASCAR Xfinity Series Results

A collection of NASCAR's second-tier competitive race series results from the first race in 1982 through the completion of the 2023 season. The series has undergone several name changes, including the Budweiser Late Model Sportsman Series, the NASCAR Busch Grand National Series, and the current NASCAR Xfinity Series.

xfinity_race_data

Format

A data frame with 51,672 rows and 23 variables:

season

num: Racing season (1982-2023)

race

num: Chronological race number for that season (1-35)

site

chr: Location of the race (city)

track

chr: Race track name

track_length

num: Race track length, in miles

track_type

chr: Race track surface (dirt track, paved track, road course)

finish

num: Finish position in the race

start

num: Start position in the race

driver

chr: Driver's name

manufacturer

chr: Car manufacturer (e.g., "Chevrolet", "Ford", "Toyota")

car_number

chr: Car number

owner

chr: Car owner

sponsor

chr: Car sponsor

win

num: Indicator variable for win; 1 = win, 0 = did not win

top_5

num: Indicator variable for finish position in 5th place or better; 1 = yes, 0 = no

top_10

num: Indicator variable for finish position in 10th place or better; 1 = yes, 0 = no

top_20

num: Indicator variable for finish position in 20th place or better; 1 = yes, 0 = no

laps

num: Number of laps completed in the race

laps_led

num: Number of laps led in the race

status

chr: Status at the end of the race (e.g., "Running", "Accident")

money

num: Total race earning (in dollars) not including bonus money, if available. No available data beginning with the 2016 season.

pts

num: Points earned in the race

playoff_pts

num: Playoff points earned in the race. Available beginning with the 2017 season.

Source

https://www.nascar.com/news/nascar-xfinity-series/

https://www.driveraverages.com/nascar_xfinityseries/

https://www.racing-reference.info/nascar-xfinity-series-stats/

https://en.wikipedia.org/wiki/NASCAR_Xfinity_Series

Examples

data(xfinity_race_data)